zero-shot-object-detection/requirements.txt
tqdm
opencv-python
transformers
numpy
scikit-image
pillow
tensorflow